requirements.txt
boto3
pandas
python-dotenv
openpyxl
requests
msal